Move xxhash64 out of the crypto dir (thank you @AljoschaMeyer)
